+++ jvm/src/main/scala/com/nawforce/apexlink/rpc/HoverItem.scala
@@ -6,7 +6,11 @@ package com.nawforce.apexlink.rpc
 import com.nawforce.pkgforce.path.Location
 import io.github.shogowada.scala.jsonrpc.serializers.JSONRPCPickler.{macroRW, ReadWriter => RW}
 
-case class HoverItem(content: Option[String], location: Option[Location])
+case class HoverItem(
+  content: Option[String],
+  location: Option[Location],
+  kind: Option[String] = None
+)
 
 object HoverItem {
   implicit val rw: RW[HoverItem]        = macroRW